gamerates moved game detail page. add average of gamerates on game detail page. add credit card annotation. add password annotation.

+++ Models/Gamerates.cs
@@ -12,7 +12,7 @@ namespace VirtualGameStore.Models
         public decimal? Userid { get; set; }
         [DisplayName("Game")]
         public decimal? Gameid { get; set; }
-        [Range(typeof(Int32), "0", "100", ErrorMessage = "{0} must be a number between {1} and {2}.")]
+        [Range(typeof(decimal), "0", "100", ErrorMessage = "{0} must be a number between {1} and {2}.")]
         public decimal? Rates { get; set; }
         [MaxLength(200)]
         public string Description { get; set; }
